feat : viewer 페이지 URL 정리리
